tests: update evalute retriever tests
